% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap.R
\name{bootdiff}
\alias{bootdiff}
\title{Difference between Two Groups with Bootstrap Confidence Intervals}
\usage{
bootdiff(data, x, y, control_group, test_group, paired, ci = 0.95,
  reps = 5000, func = mean)
}
\arguments{
\item{data}{A data.frame or tibble.}

\item{x, y}{Columns in \code{data}.}

\item{control_group, test_group}{Factors or strings in the \code{x} columns.
These must be quoted (ie surrounded by quotation marks), and will be used to
identify the control group and the test group. Any NaNs will be removed with
\code{na.omit}.}

\item{paired}{boolean. If TRUE, the two groups are treated as paired
samples. The \code{control_group} group is treated as pre-intervention
and the \code{test_group} group is considered post-intervention.}

\item{ci}{float, default 0.95. The level of the confidence intervals
produced. The default \code{ci = 0.95} produces 95\% CIs.}

\item{reps}{integer, default 5000. The number of bootstrap resamples that
will be generated.}

\item{func}{function, default mean. This function will be applied to
\code{control} and \code{test} individually, and the difference will be
saved as a single bootstrap resample.}
}
\value{
A list of the class "boot.diff", with the following 9 elements:

\item{func}{
The \code{func} passed to \code{bootdiff}.
}
\item{difference}{
The difference between the two groups;
effectively \code{func(test) - func(control)}.
}
\item{ci}{
The \code{ci} passed to the \code{bootdiff}.
}
\item{bca_ci_low}{
The lower limit of the Bias Corrected and Accelerated bootstrap confidence
interval.}

\item{bca_ci_high}{
  The upper limit of the Bias Corrected and Accelerated bootstrap confidence
  interval.}

\item{pct_ci_low}{
  The upper limit of the percentile bootstrap confidence interval.
  }
\item{pct_ci_high}{
  The lower limit of the percentile bootstrap confidence interval.
  }

\item{bootstraps}{
  The array of bootstrap resamples generated.}
}
\description{
\code{bootdiff} applies a summary function (\code{func}, default
\code{mean}) to \code{control} and \code{test}. The difference between
\code{func(test)} and \code{func(control)} is computed, and a bootstrap
confidence interval is constructed for the difference.
}
\section{References}{

DiCiccio, Thomas J., and Bradley Efron. Bootstrap Confidence Intervals.
  Statistical Science: vol. 11, no. 3, 1996, pp. 189–228,

  \url{http://www.jstor.org/stable/2246110.}


Efron, Bradley, and R. J. Tibshirani. An Introduction to the Bootstrap.
  CRC Press, 1994.

  \url{https://www.crcpress.com/An-Introduction-to-the-Bootstrap/Efron-Tibshirani/p/book/9780412042317}
}

\examples{
# Performing unpaired (two independent groups) analysis.
unpaired_mean_diff <- bootdiff(iris, Species, Petal.Width,
                               "setosa", "versicolor", paired = FALSE)

# Display the results in a user-friendly format.
unpaired_mean_diff

# Plot the bootstrap resamples as a histogram, along with the CI.
plot(unpaired_mean_diff)


# Performing paired analysis.
paired_mean_diff <- bootdiff(iris, Species, Petal.Width,
                             "setosa", "versicolor", paired = FALSE)


# Computing the median difference.
unpaired_median_diff <- bootdiff(iris, Species, Petal.Width,
                                 "setosa", "versicolor", paired = FALSE,
                                 func = median)


# Producing a 90\\\% CI instead of 95\\\%
unpaired_mean_diff_90_ci <- bootdiff(iris, Species, Petal.Width,
                                     "setosa", "versicolor", paired = FALSE,
                                     ci = 0.90)


# Constructing the confidence intervals on 10000 bootstrap resamples.
unpaired_mean_diff_n10000 <- bootdiff(iris, Species, Petal.Width,
                                      "setosa", "versicolor", paired = FALSE,
                                      reps = 10000)

}
