% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_effectsize.R
\name{cliffs_delta}
\alias{cliffs_delta}
\title{Compute Cliff's delta}
\usage{
cliffs_delta(x, ..., ci = 95, reps = 5000, seed = 12345)
}
\arguments{
\item{x}{A dabest_proto object}

\item{ci}{float, default 95. The level of the confidence intervals produced.
The default \code{ci = 95} produces 95\% CIs.}

\item{reps}{integer, default 5000. The number of bootstrap resamples that
will be generated.}

\item{seed}{integer, default 12345. This specifies the seed used to set the
random number generator. Setting a seed ensures that the bootstrap confidence
intervals for the same data will remain stable over separate runs/calls of
this function. See \link{set.seed} for more details.}
}
\value{
Difference of means for treatment vs. control.
}
\description{
Compute Cliff's delta
}
