% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_effectsize.R
\name{cohen_d_standardizers}
\alias{cohen_d_standardizers}
\title{Compute the standardizers for Cohen's d}
\usage{
cohen_d_standardizers(x1, x2)
}
\arguments{
\item{x1}{A vector of values.}

\item{x2}{Another vector of values.}
}
\value{
Hedges' correction for the g of x1 and x2.
}
\description{
Compute the standardizers for Cohen's d
}
