% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_effectsize.R
\name{cohens_d}
\alias{cohens_d}
\title{Compute Cohen's d}
\usage{
cohens_d(x, ..., ci = 95, reps = 5000, seed = 12345)
}
\arguments{
\item{ci}{float, default 95. The level of the confidence intervals produced.
The default \code{ci = 95} produces 95\% CIs.}

\item{reps}{integer, default 5000. The number of bootstrap resamples that
will be generated.}

\item{seed}{integer, default 12345. This specifies the seed used to set the
random number generator. Setting a seed ensures that the bootstrap confidence
intervals for the same data will remain stable over separate runs/calls of
this function. See \link{set.seed} for more details.}

\item{.data}{A dabest_proto object}
}
\value{
Cohen's d.
}
\description{
Compute Cohen's d
}
