% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dev.R
\name{dabest.plot.dev}
\alias{dabest.plot.dev}
\title{Estimation Plot (DEV VERSION)}
\usage{
dabest.plot.dev(.data, x, y, idx, color.col = NULL, paired = FALSE,
  func = mean, ci = 0.95, reps = 5000, slopegraph = TRUE,
  float.contrast = NULL, color.column = NULL, palette = "Set1",
  theme = theme_classic(), tick.fontsize = 12,
  axes.title.fontsize = 15, swarm.label = NULL, es.label = NULL,
  esmarker.size = 3, rawplot.type = c("beeswarm", "sinaplot"),
  beeswarm.params = NULL, sinaplot.params = NULL,
  slopegraph.params = NULL)
}
\arguments{
\item{.data}{A tidy data.frame or tibble.}

\item{x, y}{Columns in the data.frame.}

\item{idx}{Categories in the column \code{x}.}

\item{paired}{boolean. If TRUE, the two groups are treated as paired
samples. The group first passed to \code{idx} is treated as pre-intervention
while the second group is considered post-intervention.}

\item{func}{function, default mean. This function will be applied to
the groups in \code{idx} individually, and the difference will be
saved as a single bootstrap resample.}

\item{ci}{float, default 0.95. The level of the confidence intervals
produced. The default \code{ci = 0.95} produces 95\% CIs.}

\item{reps}{integer, default 5000. The number of bootstrap resamples that
will be generated.}

\item{slopegraph}{boolean, default TRUE.}

\item{float.contrast}{default NULL}

\item{color.column}{default NULL}

\item{palette}{default "Set1"}

\item{theme}{default theme_classic()}

\item{tick.fontsize}{default 12}

\item{axes.title.fontsize}{default 15}

\item{swarm.label}{default NULL}

\item{es.label}{default NULL}

\item{esmarker.size}{default 3}

\item{rawplot.type}{default "beeswarm. Accepts any value found in
c("beeswarm", "sinaplot")}

\item{beeswarm.params}{default NULL. Supply list of \code{keyword = value}
pairs to \code{ggbeeswarm::geom_beeswarm()}.}

\item{sinaplot.params}{default NULL. Supply list of \code{keyword = value}
pairs to \code{ggforce::geom_sina()}.}

\item{slopegraph.params}{default NULL. Supply list of \code{keyword = value}
pairs to \code{ggplot2::geom_line()}. This controls the appearance of the
lines plotted for a paired slopegraph.}
}
\value{
a \code{ggplot} object.
}
\description{
An estimation plot has two key features.
\enumerate{
  \item{It presents all datapoints as a
  \href{https://github.com/eclarke/ggbeeswarm#introduction}{swarmplot} or
  \href{https://cran.r-project.org/web/packages/sinaplot/vignettes/SinaPlot.html}{sinaplot},
  which orders each point to display the underlying distribution.}
  \item{It presents the effect size as a bootstrap 95 percent confidence
        interval on a separate but aligned axes.}
}
Estimation plots emerge from estimation statistics, an inutitive framework
that avoids the pitfalls of significance testing. It uses familiar
statistical concepts: means, mean differences, and error bars.
More importantly, it focuses on the effect size of one's
experiment/intervention, as opposed to a false dichotomy engendered
by \emph{P} values.
}
\section{References}{

Moving beyond P values: Everyday data analysis with estimation plots.
Joses Ho, Tayfun Tumkaya, Sameer Aryal, Hyungwon Choi, Adam Claridge-Chang (2018)

\url{https://doi.org/10.1101/377978}
}

\examples{
dabest.plot.dev(iris, x = Species, y = Petal.Width,
                idx = c("setosa", "versicolor"))

}
