% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{dabest.plot}
\alias{dabest.plot}
\title{Estimation Plot}
\usage{
dabest.plot(data, x, y, idx, fun = mean, reps = 5000)
}
\arguments{
\item{data}{A tidy data.frame.}

\item{x, y}{Columns in the data.frame.}

\item{idx}{Categories in the column \code{x}.}

\item{fun}{function, default mean. This function will be applied to
the groups in \code{idx} individually, and the difference will be
saved as a single bootstrap resample.}

\item{reps}{integer, default 5000. The number of bootstrap resamples that
will be generated.}

\item{paired}{boolean. If TRUE, the two groups are treated as paired
samples. The group first passed to \code{idx} is treated as pre-intervention
while the second group is considered post-intervention.}

\item{ci}{float, default 0.95. The level of the confidence intervals
produced. The default \code{ci = 0.95} produces 95 percent CIs.}
}
\value{
a \code{ggplot} object.
}
\description{
An estimation plot has two key features.
\enumerate{
  \item{It presents all datapoints as a
  \href{https://github.com/eclarke/ggbeeswarm#introduction}{swarmplot} or
  \href{https://cran.r-project.org/web/packages/sinaplot/vignettes/SinaPlot.html}{sinaplot},
  which orders each point to display the underlying distribution.}
  \item{It presents the effect size as a bootstrap 95 percent confidence
        interval on a separate but aligned axes.}
}
Estimation plots emerge from estimation statistics, an inutitive framework
that avoids the pitfalls of significance testing. It uses familiar
statistical concepts: means, mean differences, and error bars.
More importantly, it focuses on the effect size of one's
experiment/intervention, as opposed to a false dichotomy engendered
by \emph{P} values.
}
\section{References}{

Moving beyond P values: Everyday data analysis with estimation plots.
Joses Ho, Tayfun Tumkaya, Sameer Aryal, Hyungwon Choi, Adam Claridge-Chang (2018)

\url{https://doi.org/10.1101/377978}
}

\examples{
# Coming soon!

}
