% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_base.R
\name{dabest_proto}
\alias{dabest_proto}
\title{Prepare Data for Analysis with dabestr}
\usage{
dabest_proto(.data, x, y, idx, paired = FALSE, id.column = NULL)
}
\arguments{
\item{.data}{A data.frame or tibble.}

\item{x, y}{Columns in \code{.data}.}

\item{idx}{A vector containing factors or strings in the \code{x} columns.
These must be quoted (ie. surrounded by quotation marks). The first element
will be the control group, so all differences will be computed for every
other group and this first group.}

\item{paired}{boolean, default FALSE. If TRUE, the two groups are treated as
paired samples. The \code{control_group} group is treated as
pre-intervention and the \code{test_group} group is considered
post-intervention.}

\item{id.column, }{default NULL. A column name indicating the identity of the
datapoint if the data is paired. This must be supplied if paired is TRUE.}
}
\value{
A list with 7 elements: \code{data}, \code{x}, \code{y}, \code{idx},
 and \code{id.column}.

  \code{data}, \code{x}, \code{y}, \code{id.column}, and \code{idx} are the
  same keywords supplied to \code{dabest} as noted above. \cr \cr \code{x}
  and \code{y} are quoted variables for tidy evaluation by \code{plot}.
}
\description{
Prepare Data for Analysis with dabestr
}
