% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabestr.R
\docType{package}
\name{dabestr}
\alias{dabestr}
\alias{dabestr-package}
\title{dabestr: A package for producing estimation plots.}
\description{
The dabestr package provides a function to generate estimation plots
\code{dabest.plot} and a function to construct bootstrap confidence intervals
for two-group differences \code{bootdiff}.
}
