% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{deltadelta_data}
\alias{deltadelta_data}
\title{Data to produce a delta2 Dabest plot}
\format{
A data frame with 40 rows and 5 variables:
\describe{
\item{Genotype}{Genotype of each observation}
\item{ID}{Identity of each observation}
\item{Rep}{Rep of each observation}
\item{Treatment}{Which treatment method was used}
\item{Measurement}{Measurement value}
}
}
\usage{
deltadelta_data
}
\description{
Contains 2 Genotype groups and 2 Treatment groups.
}
\examples{
data(deltadelta_data) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{internal}
