% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{get_tick_labels}
\alias{get_tick_labels}
\title{Get tick labels from a ggplot object.}
\usage{
get_tick_labels(plot.obj, axes)
}
\arguments{
\item{plot.obj}{A \code{ggplot} object.}

\item{axes}{string, either "x" or "y".}
}
\description{
Requires ggplot2 v3.0.0.
}
\examples{
my.plot <- ggplot(iris, aes(Species, Petal.Width)) + geom_boxplot()
my.plot.yticks <- get_tick_labels(my.plot, axes="y")
my.plot.xticks <- get_tick_labels(my.plot, axes="x")

}
