% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsize.R
\name{hedges_correction}
\alias{hedges_correction}
\title{Returns the exact Hedges' correction factor for Cohen's d.}
\usage{
hedges_correction(x1, x2)
}
\arguments{
\item{x1}{A vector of values.}

\item{x2}{Another vector of values.}
}
\value{
Hedges' correction for the g of x1 and x2.
}
\description{
Returns the exact Hedges' correction factor for Cohen's d.
}
