% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsat_scores}
\alias{lsat_scores}
\title{LSAT Scores.}
\format{A list with two elements: \code{pretest} and \code{posttest}.}
\source{
This was obtained from xxxyyy
}
\usage{
lsat_scores
}
\description{
Describe the experiment and the dataset.
}
\keyword{datasets}
