% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{max_nchar_ticks}
\alias{max_nchar_ticks}
\title{Get the longest character string from a list.}
\usage{
max_nchar_ticks(tick_list)
}
\arguments{
\item{tick_list}{A list of strings.}
}
\description{
Get the longest character string from a list.
}
\examples{
my.plot <- ggplot(iris, aes(Species, Petal.Width)) + geom_boxplot()
my.plot.yticks <- get_tick_labels(my.plot, axes="y")
max_ticklength   <- max_nchar_ticks(my.plot.yticks)

}
