% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{minimeta_data}
\alias{minimeta_data}
\title{Data to produce a mini-meta Dabest plot}
\format{
A data frame with 120 rows and 5 variables:
\describe{
\item{Gender}{Gender of each observation}
\item{ID}{Identity of each observation}
\item{Group}{Which control group or test it is}
\item{Measurement}{Measurement value}
}
}
\usage{
minimeta_data
}
\description{
Contains 3 Control Samples and 3 Test Samples.
}
\examples{
data(minimeta_data) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{internal}
