% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{non_proportional_data}
\alias{non_proportional_data}
\title{Non-proportional data for Estimation plots.}
\format{
A data frame with 180 rows and 4 variables:
\describe{
\item{Gender}{Gender of each observation}
\item{ID}{Identity of each observation}
\item{Group}{Which control group or test it is}
\item{Measurement}{Measurement value}
}
}
\usage{
non_proportional_data
}
\description{
Contains 3 Control Samples and 6 Test Samples.
}
\examples{
data(non_proportional_data) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{internal}
