% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.dabest}
\alias{plot.dabest}
\title{Estimation Plot}
\usage{
\method{plot}{dabest}(dabest.object, color.column = NULL,
  palette = "Set1", float.contrast = TRUE, slopegraph = TRUE,
  theme = theme_classic(), tick.fontsize = 11,
  axes.title.fontsize = 14, swarm.label = NULL, es.label = NULL,
  rawmarker.size = 2, esmarker.size = 4,
  rawplot.type = c("swarmplot", "sinaplot"), swarmplot.params = NULL,
  sinaplot.params = NULL, slopegraph.params = NULL)
}
\arguments{
\item{dabest.object}{A \code{dabest} object, generated by the function
of the same name.}

\item{color.column}{default NULL}

\item{palette}{default "Set1". Accepts any one of the ggplot2 palettes.
See the "Palettes" section ins \link{scale_color_brewer}.}

\item{float.contrast}{default TRUE.  If \code{idx} in the \code{dabest}
object contains only 2 groups, \code{float.contrast = TRUE} will plot the
effect size and the bootstrap confidence interval in a horizontally-aligned
axes (also known as a \href{https://bit.ly/2NhqUAb}{Gardner-Altman plot.})}

\item{slopegraph}{boolean, default TRUE. If the \code{dabest} object contains
paired comparisons, \code{slopegraph = TRUE} will plot the rawdata as a
\href{http://charliepark.org/slopegraphs/}{Tufte slopegraph}.}

\item{theme}{default theme_classic()}

\item{tick.fontsize}{default 12}

\item{axes.title.fontsize}{default 15}

\item{swarm.label}{default NULL}

\item{es.label}{default NULL}

\item{rawmarker.size}{default 2}

\item{esmarker.size}{default 4}

\item{rawplot.type}{default "beeswarm. Accepts any value found in
c("beeswarm", "sinaplot")}

\item{swarmplot.params}{default NULL. Supply list of
\code{keyword = value} pairs to \code{ggbeeswarm::geom_beeswarm()}
(if float.contrast = TRUE) or \code{ggbeeswarm::geom_quasirandom()}
(if float.contrast = FALSE).}

\item{sinaplot.params}{default NULL. Supply list of
\code{keyword = value} pairs to \code{ggforce::geom_sina()}.}

\item{slopegraph.params}{default NULL. Supply list of
\code{keyword = value} pairs to \code{ggplot2::geom_line()}. This
controls the appearance of the lines plotted for a paired slopegraph.}
}
\value{
A \code{ggplot} object.
}
\description{
An estimation plot has two key features.
\enumerate{
  \item{It presents all datapoints as a
  \href{https://github.com/eclarke/ggbeeswarm#introduction}{swarmplot} or
  \href{https://cran.r-project.org/web/packages/sinaplot/vignettes/SinaPlot.html}{sinaplot},
  which orders each point to display the underlying distribution.}
  \item{It presents the effect size as a bootstrap 95 percent confidence
        interval on a separate but aligned axes.}
}
Estimation plots emerge from estimation statistics, an inutitive framework
that avoids the pitfalls of significance testing. It uses familiar
statistical concepts: means, mean differences, and error bars.
More importantly, it focuses on the effect size of one's
experiment/intervention, as opposed to a false dichotomy engendered
by \emph{P} values.
This function takes the output of the \code{\link{dabest}} function
and produces an estimation plot.
}
\section{References}{

Moving beyond P values: Everyday data analysis with estimation plots. (2018)
Joses Ho, Tayfun Tumkaya, Sameer Aryal, Hyungwon Choi, Adam Claridge-Chang

\url{https://doi.org/10.1101/377978}
}

\examples{
paired_mean_diff <- dabest(
                      iris, Species, Petal.Width,
                      idx = c("setosa", "versicolor", "virginica"),
                      paired = FALSE)
plot(paired_mean_diff, color.col = Species)

}
\seealso{
The \code{\link{dabest}} function.
}
