% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_plot.R
\name{plot.dabest_effsize_proto}
\alias{plot.dabest_effsize_proto}
\title{Create an estimation plot!}
\usage{
\method{plot}{dabest_effsize_proto}(
  x,
  ...,
  color.column = NULL,
  palette = "Set1",
  float.contrast = TRUE,
  slopegraph = TRUE,
  group.summaries = "mean_sd",
  rawplot.type = c("swarmplot", "sinaplot"),
  rawplot.ylim = NULL,
  rawplot.ylabel = NULL,
  rawplot.markersize = 2,
  rawplot.groupwidth = 0.3,
  effsize.ylim = NULL,
  effsize.ylabel = NULL,
  effsize.markersize = 4,
  theme = ggplot2::theme_classic(),
  tick.fontsize = 11,
  axes.title.fontsize = 14,
  swarmplot.params = NULL,
  sinaplot.params = NULL,
  slopegraph.params = NULL
)
}
\description{
Create an estimation plot!
}
