% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.dabest}
\alias{print.dabest}
\title{Print a `dabest` object}
\usage{
\method{print}{dabest}(dabest.object, signif_digits = 3)
}
\arguments{
\item{signif_digit}{integer, default 3. All numerical values in the printout
will be rounded to this many significant digits.}
}
\description{
Print a `dabest` object
}
