% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_base.R
\name{print.dabest_proto}
\alias{print.dabest_proto}
\title{Print a `dabest_proto` object}
\usage{
\method{print}{dabest_proto}(dabest.object, ...)
}
\arguments{
\item{dabest.object}{A \code{proto_dabest} object, generated by the function of the same name.}

\item{...}{Signature for S3 generic function.}
}
\value{
A summary of the experimental designs.
}
\description{
Print a `dabest_proto` object
}
\examples{
# Performing unpaired (two independent groups) analysis.
unpaired_mean_diff <- dabest(iris, Species, Petal.Width,
                             idx = c("setosa", "versicolor"),
                             paired = FALSE)

# Display the results in a user-friendly format.
print(unpaired_mean_diff)

}
