% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000_data.R
\docType{data}
\name{proportional_data}
\alias{proportional_data}
\title{Numerical Binary data for Proportion Plots}
\format{
A data frame with 400 rows and 4 variables:
\describe{
\item{Gender}{Gender of each observation}
\item{ID}{Identity of each observation}
\item{Group}{Which control group or test it is}
\item{Success}{1 (Success) or 0 (Failure)}
}
}
\usage{
proportional_data
}
\description{
Contains 3 Control Samples and 7 Test Samples.
}
\examples{
data(proportional_data) # Lazy loading. Data becomes visible as soon as it is loaded
}
\keyword{internal}
