% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wellbeing_ind}
\alias{wellbeing_ind}
\title{Wellbeing Scores (2 independent groups).}
\format{A list with two elements: \code{pretest} and \code{posttest}.}
\source{
This was obtained from xxxyyy
}
\usage{
wellbeing_ind
}
\description{
Describe the experiment and the dataset.
}
\keyword{datasets}
