% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wellbeing_paired}
\alias{wellbeing_paired}
\title{Wellbeing Scores (Before and after design).}
\format{A list with two elements: \code{pretest} and \code{posttest}.}
\source{
This was obtained from xxxyyy
}
\usage{
wellbeing_paired
}
\description{
Describe the experiment and the dataset.
}
\keyword{datasets}
