% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_functions.R
\name{bootdiff}
\alias{bootdiff}
\title{Difference between two groups with bootstrap confidence intervals}
\usage{
bootdiff(data, control, test, paired, ci = 0.95, reps = 5000,
  func = mean)
}
\arguments{
\item{data}{A data.frame or list.}

\item{control}{A column or item in `data`. This is the control group.}

\item{test}{A column or item in `data`. This is the test group.}

\item{paired}{boolean. If TRUE, the two groups are treated as paired
samples. The \code{control} group is treated as pre-intervention
and the \code{control} group is considered post-intervention.}

\item{ci}{float, default 0.95. The level of the confidence intervals
produced. The default \code{ci = 0.95} produces 95 percent CIs.}

\item{reps}{integer, default 5000. The number of bootstrap resamples that
will be generated.}

\item{func}{function, default mean. This function will be applied to
\code{control} and \code{test} individually, and the difference will be
saved as a single bootstrap resample.}
}
\value{
A list with the following 8 elements:

\item{func}{
The \code{func} passed to \code{bootdiff}.
}
\item{difference}{
The difference between the two groups;
effectively \code{func(test) - func(control)}.
}
\item{ci}{
The \code{ci} passed to the \code{bootdiff}.
}
\item{bca_ci_low}{
The lower limit of the Bias Corrected and Accelerated bootstrap confidence
interval.}

\item{bca_ci_high}{
  The upper limit of the Bias Corrected and Accelerated bootstrap confidence
  interval.}

\item{pct_ci_low}{
  The upper limit of the percentile bootstrap confidence interval.
  }
\item{pct_ci_high}{
  The lower limit of the percentile bootstrap confidence interval.
  }

\item{bootstraps}{
  The array of bootstrap resamples generated.}
}
\description{
\code{bootdiff} returns the difference between 2 groups.
}
\section{References}{

1. What is bootstrap resampling?
2. What are BCa confidence intervals?
}

\examples{
# Use the preloaded data.

}
\seealso{
\url{https://www.google.com}
}
