% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effectsize.R
\name{cohen_d_standardizers}
\alias{cohen_d_standardizers}
\title{Compute the standardizers for Cohen's d}
\usage{
cohen_d_standardizers(x1, x2)
}
\arguments{
\item{x1}{A vector of values.}

\item{x2}{Another vector of values.}
}
\value{
Named list for pooled and average standardizers.
Use the pooled one for unpaired Cohens' d, and the average one
for paired Cohens'd d.
}
\description{
Compute the standardizers for Cohen's d
}
