% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabestr.R
\docType{package}
\name{dabestr}
\alias{dabestr}
\title{dabestr: A package for producing estimation plots.}
\description{
The dabestr package provides functiond to construct bootstrap confidence
intervals for differences between groups for a range of
\link[=mean_diff]{effect sizes} and a function to
generate \link[=plot.dabest_effsize]{estimation plots}.
}
