% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dabestr.R
\docType{package}
\name{dabestr}
\alias{dabestr}
\alias{dabestr-package}
\title{dabestr: A package for producing estimation plots.}
\description{
The dabestr package provides a function to construct bootstrap confidence
intervals for differences between groups \link{dabest} and a function to
generate estimation plots \link{plot.dabest}.
}
