% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lsat_scores}
\alias{lsat_scores}
\title{LSAT Scores.}
\format{
A list with two elements: \code{pretrain} and \code{posttrain}.
}
\source{
Thomason, N. R., Adajian, T., Barnett, A. E., Boucher, S.,
van der Brugge, E., Campbell, J., Knorpp, W., Lempert, R.,
Lengbeyer, L., Mandel, D. R., Rider, Y., van Gelder, T., & Wilkins, J. (2014).
\emph{Critical thinking final report}. The University of Melbourne,
N66001-12-C-2004.
}
\usage{
lsat_scores
}
\description{
This dataset is taken from Thomason et al. (2014) who studied a group of
12 students who underwent a course in critical thinking. Their scores on
the Logical Reasoning section of the Law School Aptitude Test (LSAT) were
assesed before and after training.
}
\details{
It is also found in Chapter 8 "The Paired Design"
(pp 195 - 199) of Introduction to the New Statistics
(Routledge, 2017), by Geoff Cumming and Robert Calin-Jageman.
}
\keyword{datasets}
