% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{print.dabest}
\alias{print.dabest}
\title{Print a `dabest` object}
\usage{
\method{print}{dabest}(x, ...)
}
\arguments{
\item{x}{A \code{\link{dabest}} object, generated by the function of the same name.}

\item{...}{S3 signature for generic plot function.}
}
\value{
A summary of the experimental designs.
}
\description{
Print a `dabest` object
}
\examples{
# Performing unpaired (two independent groups) analysis.
unpaired_mean_diff <- dabest(iris, Species, Petal.Width,
                             idx = c("setosa", "versicolor"),
                             paired = FALSE)

# Display the results in a user-friendly format.
print(unpaired_mean_diff)

}
