% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prototype_effectsize.R
\name{print.dabest_effsize_proto}
\alias{print.dabest_effsize_proto}
\title{Print a `dabest_effsize_proto` object}
\usage{
\method{print}{dabest_effsize_proto}(x, ..., signif_digits = 3)
}
\arguments{
\item{...}{Signature for S3 generic function.}

\item{dabest.effsize}{A \code{dabest_effsize_proto} object, generated by \code{compute.effect.size}.}
}
\value{
A summary of the experimental designs.
}
\description{
Print a `dabest_effsize_proto` object
}
\examples{
# Performing unpaired (two independent groups) analysis.
unpaired_mean_diff <- dabest(iris, Species, Petal.Width,
                             idx = c("setosa", "versicolor"),
                             paired = FALSE) \%>\%
                     compute.effect.siz

# Display the results in a user-friendly format.
print(unpaired_mean_diff)

}
