% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{transcription_scores}
\alias{transcription_scores}
\title{Transcription Scores.}
\format{
A list with two elements: \code{pen} and \code{laptop}.
}
\source{
Mueller, Pam A., and Daniel M. Oppenheimer. “The Pen Is Mightier
than the Keyboard: Advantages of Longhand over Laptop Note Taking.”
Psychological Science, vol. 25, no. 6, June 2014, pp. 1159–68,
doi:10.1177/0956797614524581.
}
\usage{
transcription_scores
}
\description{
This dataset is taken from Mueller and Oppenheimer (2014), comparing the
percentage of notes that was verbatim transcribed during a lecture by two
independent groups of students: one using pen and paper, and one using
laptops.
}
\details{
It is also found in Chapter 7 "The Independent Groups Design"
(pp 160 - 166) of Introduction to the New Statistics
(Routledge, 2017), by Geoff Cumming and Robert Calin-Jageman.
}
\keyword{datasets}
