% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wellbeing_paired}
\alias{wellbeing_paired}
\title{Wellbeing Scores (Before and after design).}
\format{
A list with two elements: \code{before} and \code{after}.
}
\source{
Cumming, G. Understanding the New Statistics: Effect Sizes,
Confidence Intervals, and Meta-Analysis. Routledge 2012.
https://books.google.com/books?id=AVBDYgEACAAJ
}
\usage{
wellbeing_paired
}
\description{
This is a synthetic dataset from Geoff Cumming. A single group of 10 students
first completed a well-being questionnaire (\code{before}), spent the
afternoon reading in the botanical gardens, then gave well-being scores once
again (\code{after}).
}
\details{
It is found in Chapter 11 "The Paired Design" (page 291) of
Understanding the New Statistics (Routledge, 2012) by Geoff Cumming.
}
\keyword{datasets}
